+++ src/Concerto/TestBundle/Command/StartForkerCommand.php
@@ -15,14 +15,16 @@ class StartForkerCommand extends Command
     private $testRunnerSettings;
     private $doctrine;
     private $sessionRunnerService;
+    private $projectDir;
 
-    public function __construct($testRunnerSettings, RegistryInterface $doctrine, ASessionRunnerService $sessionRunnerService)
+    public function __construct($testRunnerSettings, RegistryInterface $doctrine, ASessionRunnerService $sessionRunnerService, $projectDir)
     {
         parent::__construct();
 
         $this->doctrine = $doctrine;
         $this->testRunnerSettings = $testRunnerSettings;
         $this->sessionRunnerService = $sessionRunnerService;
+        $this->projectDir = $projectDir;
     }
 
     protected function configure()
@@ -32,45 +34,45 @@ class StartForkerCommand extends Command
 
     private function getCommand()
     {
-        $forkerPath = realpath(dirname(__FILE__) . "/../Resources/R/forker.R");
-        $logPath = realpath(dirname(__FILE__) . "/../Resources/R") . "/forker.log";
-        $fifoPath = $this->sessionRunnerService->getFifoDir();
-        $publicDir = $this->sessionRunnerService->getPublicDirPath();
-        $connection = json_encode($this->sessionRunnerService->getConnection());
-        $platformUrl = $this->sessionRunnerService->getPlatformUrl();
-        $appUrl = $this->sessionRunnerService->getAppUrl();
-        $maxExecTime = $this->testRunnerSettings["max_execution_time"];
-        $maxIdleTime = $this->testRunnerSettings["max_idle_time"];
-        $keepAliveToleranceTime = $this->testRunnerSettings["keep_alive_tolerance_time"];
+        $forkerPath = "{$this->projectDir}/src/Concerto/TestBundle/Resources/R/forker.R";
+        $logPath = "{$this->projectDir}/var/logs/forker.log";
 
-        $cmd = "nohup " . $this->testRunnerSettings["rscript_exec"] . " --no-save --no-restore --quiet "
-            . "'$forkerPath' "
-            . "'$fifoPath' "
-            . "'$publicDir' "
-            . "'$platformUrl' "
-            . "'$appUrl' "
-            . "'$connection' "
-            . "$maxExecTime "
-            . "$maxIdleTime "
-            . "$keepAliveToleranceTime "
-            . ">> "
-            . "'" . $logPath . "' "
-            . "2>&1 & echo $!";
-        return $cmd;
+        return "nohup {$this->testRunnerSettings["rscript_exec"]} --no-save --no-restore --quiet '$forkerPath' >> '$logPath' 2>&1 & echo $!";
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $output->writeln("starting forker...");
 
+        $fifoPath = $this->sessionRunnerService->getFifoDir();
+        $publicDir = $this->sessionRunnerService->getPublicDirPath();
+        $dbConnection = json_encode($this->sessionRunnerService->getDbConnectionParams());
+        $platformUrl = $this->sessionRunnerService->getPlatformUrl();
+        $appUrl = $this->sessionRunnerService->getAppUrl();
+        $maxExecTime = $this->testRunnerSettings["max_execution_time"];
+        $maxIdleTime = $this->testRunnerSettings["max_idle_time"];
+        $keepAliveToleranceTime = $this->testRunnerSettings["keep_alive_tolerance_time"];
+        $sessionStorage = $this->testRunnerSettings["session_storage"];
+        $redisConnection = json_encode($this->sessionRunnerService->getRedisConnectionParams());
+
         $cmd = $this->getCommand();
         $process = new Process($cmd);
-        $env = array(
+
+        $env = [
+            "CONCERTO_R_APP_URL" => $appUrl,
+            "CONCERTO_R_DB_CONNECTION" => $dbConnection,
+            "CONCERTO_R_FIFO_PATH" => $fifoPath,
+            "CONCERTO_R_MAX_EXEC_TIME" => $maxExecTime,
+            "CONCERTO_R_MAX_IDLE_TIME" => $maxIdleTime,
+            "CONCERTO_R_KEEP_ALIVE_TOLERANCE_TIME" => $keepAliveToleranceTime,
+            "CONCERTO_R_PLATFORM_URL" => $platformUrl,
+            "CONCERTO_R_PUBLIC_DIR" => $publicDir,
+            "CONCERTO_R_REDIS_CONNECTION" => $redisConnection,
+            "CONCERTO_R_SESSION_STORAGE" => $sessionStorage,
             "R_GC_MEM_GROW" => 0
-        );
-        if ($this->testRunnerSettings["r_environ_path"] != null) {
-            $env["R_ENVIRON"] = $this->testRunnerSettings["r_environ_path"];
-        }
+        ];
+        if ($this->testRunnerSettings["r_environ_path"] != null) $env["R_ENVIRON"] = $this->testRunnerSettings["r_environ_path"];
+
         $process->setEnv($env);
         $process->mustRun();
         if ($process->getExitCode() == 0) {