+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.redis.connect.R
@@ -1,0 +1,12 @@
+concerto.redis.connect = function(host, port, password){
+    concerto.log("connecting with redis...")
+
+    redisPass = NULL
+    if(password != "") { redisPass = password }
+
+    hiredis(
+        host = host,
+        port = port,
+        password = redisPass
+    )
+}