+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.table.lastInsertId.R
@@ -2,7 +2,7 @@ concerto.table.lastInsertId <-
 function(connection = NULL){
   id = NULL
   if(is.null(connection)) { connection = concerto$connection }
-  if(concerto$connectionParams$driver == "pdo_sqlsrv") {
+  if(concerto$dbConnectionParams$driver == "pdo_sqlsrv") {
     id = concerto$sqlsrv_last_insert_id
   } else {
     id = dbGetQuery(connection, "SELECT LAST_INSERT_ID();")[1,1]