+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * protected session files
 * JWT token for session ownership and protected global/session file access
 * data table column length can be set from panel (length of string, precision and scale for decimal)
+* redis integration for session storage
 
 #### Starter content